Added "CrosshairTextRange" key.
Additionally to "CrosshairText" key, there is also "CrosshairTextRange" now.
By default (unset) it's 9999, meaning that if "CrosshairText" is set and we are looking at the entity and are within 9999 units, the text is displayed. If "CrosshairTextRange" would be set to 1024. Then the text will only show if we are aiming at the entity within 1024 Units.